move /API/pic_email.py to /
